% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wavemat.R
\name{Wavmat}
\alias{Wavmat}
\title{Construct Orthogonal Discrete Wavelet Transformation Matrix}
\usage{
Wavmat(h, N, k, shift)
}
\arguments{
\item{h}{Low-pass filter vector.}

\item{N}{Size of the matrix.}

\item{k}{Depth of the wavelet transformation.}

\item{shift}{Integer shifts for wavelet transformation.}
}
\value{
Wavmat transformation matrix.
}
\description{
Construct Orthogonal Discrete Wavelet Transformation Matrix
}
