% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WavmatWP.R
\name{WavmatWP}
\alias{WavmatWP}
\title{Construct Wavelet Packet Transformation Matrix}
\usage{
WavmatWP(h, N, k, shift)
}
\arguments{
\item{h}{Low-pass filter vector.}

\item{N}{Size of the matrix.}

\item{k}{Depth of the wavelet packet transformation.}

\item{shift}{Integer shifts for wavelet transformation.}
}
\value{
Wavelet Packet transformation matrix.
}
\description{
Construct Wavelet Packet Transformation Matrix
}
